Add select folder logic
