+++ src/webview/create-component/pages/createComponent.tsx
@@ -5,16 +5,12 @@
 import ConstructionIcon from '@mui/icons-material/Construction';
 import FolderOpenIcon from '@mui/icons-material/FolderOpen';
 import GitHubIcon from '@mui/icons-material/GitHub';
-import { Theme, ThemeProvider, Typography } from '@mui/material';
-import { makeStyles } from '@mui/styles';
+import { PaletteMode, Theme, ThemeProvider, Typography, createTheme } from '@mui/material';
 import * as React from 'react';
-import createComponentStyle, { createVscodeTheme } from '../../common/createComponent.style';
 import { DevfileSearch } from '../../common/devfileSearch';
 import OptionCard from '../../common/optionCard';
 import { FromExistingGitRepo } from './fromExisitingGitRepo';
 import { FromLocalCodebase } from './fromLocalCodebase';
-
-const useCreateComponentStyles = makeStyles(createComponentStyle);
 
 interface VSCodeMessage {
     action: string;
@@ -24,10 +20,9 @@ interface VSCodeMessage {
 }
 
 function SelectStrategy({ setCurrentView }) {
-    const createComponentStyle = useCreateComponentStyles();
     return (
         <>
-            <div className={createComponentStyle.headerContainer}>
+            <div style={{ position: 'relative', marginTop: '5em' }}>
                 <Typography variant='h5'>Create Component</Typography>
             </div>
             <div
@@ -63,7 +58,6 @@ function SelectStrategy({ setCurrentView }) {
 }
 
 export default function CreateComponent() {
-    const createComponentStyle = useCreateComponentStyles();
     const [theme, setTheme] = React.useState<Theme>(createVscodeTheme('light'));
     const [currentView, setCurrentView] = React.useState('home');
 
@@ -88,11 +82,11 @@ export default function CreateComponent() {
             case 'home':
                 return <SelectStrategy setCurrentView={setCurrentView} />;
             case 'fromLocalCodeBase':
-                return <div className={createComponentStyle.inputContainer}>
+                return <div style={{ marginRight: '5em' }}>
                     <FromLocalCodebase setCurrentView={setCurrentView} />
                 </div>;
             case 'fromExistingGitRepo':
-                return <div className={createComponentStyle.inputContainer}>
+                return <div style={{ marginRight: '5em' }}>
                     <FromExistingGitRepo setCurrentView={setCurrentView} />
                 </div>;
             case 'devfileSearch':
@@ -102,9 +96,128 @@ export default function CreateComponent() {
 
     return (
         <ThemeProvider theme={theme}>
-            <div className={createComponentStyle.pageContainer}>
+            <div style={{ marginLeft: '10em', marginRight: '10em' }}>
                 {renderComponent()}
             </div>
         </ThemeProvider>
     );
 }
+
+const createVscodeTheme = (paletteMode: PaletteMode): Theme => {
+    const computedStyle = window.getComputedStyle(document.body);
+    return createTheme({
+        palette: {
+            mode: paletteMode,
+            primary: {
+                main: computedStyle.getPropertyValue('--vscode-button-background'),
+            },
+            error: {
+                main: computedStyle.getPropertyValue('--vscode-editorError-foreground'),
+            },
+            warning: {
+                main: computedStyle.getPropertyValue('--vscode-editorWarning-foreground'),
+            },
+            info: {
+                main: computedStyle.getPropertyValue('--vscode-editorInfo-foreground'),
+            },
+            success: {
+                main: computedStyle.getPropertyValue('--vscode-debugIcon-startForeground'),
+            },
+        },
+        typography: {
+            allVariants: {
+                fontFamily: computedStyle.getPropertyValue('--vscode-font-family'),
+            },
+        },
+        components: {
+            MuiCard: {
+                variants: [
+                    {
+                        props: {
+                            variant: 'outlined'
+                        },
+                        style: {
+                            maxWidth: '35em',
+                            backgroundColor: computedStyle.getPropertyValue('--vscode-editor-inactiveSelectionBackground')
+                        }
+                    }
+                ]
+            },
+            MuiTypography: {
+                variants: [
+                    {
+                        props: {
+                            variant: 'h5'
+                        },
+                        style: {
+                            fontSize: '2.3em',
+                            fontWeight: '650',
+                            color: computedStyle.getPropertyValue('--vscode-editor-foreground')
+                        }
+                    },
+                    {
+                        props: {
+                            variant: 'h6'
+                        },
+                        style: {
+                            fontSize: '1.5em',
+                            fontWeight: '600',
+                            color: computedStyle.getPropertyValue('--vscode-editor-foreground')
+                        }
+                    },
+                    {
+                        props: {
+                            variant: 'body1'
+                        },
+                        style: {
+                            color: computedStyle.getPropertyValue('--vscode-editor-foreground')
+                        }
+                    }
+                ]
+            },
+            MuiTextField: {
+                variants: [
+                    {
+                        props: {
+                            variant: 'outlined'
+                        },
+                        style: {
+                            id: 'outlined-basic',
+                            width: '100%',
+                            color: computedStyle.getPropertyValue('--vscode-editor-foreground'),
+                            marginBottom: '1em'
+                        }
+                    }
+                ]
+            },
+            MuiAccordion: {
+                variants: [
+                    {
+                        props: {
+                            className: 'accordion'
+                        },
+                        style: {
+                            width: '100%',
+                            color: computedStyle.getPropertyValue('--vscode-editor-foreground'),
+                            backgroundColor: computedStyle.getPropertyValue('--vscode-editor-background')
+                        }
+                    }
+                ]
+            },
+            MuiSelect: {
+                variants: [
+                    {
+                        props: {
+                            className: 'selectFolder'
+                        },
+                        style: {
+                            width: '100%',
+                            color: computedStyle.getPropertyValue('--vscode-editor-foreground'),
+                            backgroundColor: computedStyle.getPropertyValue('--vscode-editor-background')
+                        }
+                    }
+                ]
+            },
+        },
+    });
+};
